move random number generator into sys module
