+++ libs/megadrive-sys/src/lib.rs
@@ -7,6 +7,7 @@ pub mod vdp;
 pub mod ports;
 pub mod fm;
 pub mod psg;
+pub mod rng;
 
 extern "C" {
     static _data_src: *const u32;
